+++ R/box_usage_helper_functions.R
@@ -43,7 +43,7 @@ get_declared_functions <- function(xml) {
 #' @return A list of `xml_nodes` and `text`.
 get_declared_objects <- function(xml) {
   xpath_object_assignment <- "
-  //expr[LEFT_ASSIGN]/expr[1]/SYMBOL[1] |
+  //expr[LEFT_ASSIGN]/expr[1]/SYMBOL[1 and not(preceding-sibling::OP-DOLLAR)] |
   //equal_assign/expr[1]/SYMBOL[1] |
   //expr_or_assign_or_help/expr[1]/SYMBOL[1] |
   //expr[expr[1][SYMBOL_FUNCTION_CALL/text()='assign']]/expr[2]/* |